Flight Code - HAVOC 19

+++ src/havoc.cpp
@@ -116,11 +116,11 @@ void stateActions() {
       blinkLEDs();
       break;
     case FlightState::STABILIZATION:
-      if (data.gps.pos.alt % 1000 < 250) {
+      if ((int)data.gps.pos.alt % 1000 < 250) {
         data.target = targetPresets.north->getTarget(data);  
-      } else if (data.gps.pos.alt % 1000 < 500) {
+      } else if ((int)data.gps.pos.alt % 1000 < 500) {
         data.target = targetPresets.east->getTarget(data);  
-      } else if (data.gps.pos.alt % 1000 < 750) {
+      } else if ((int)data.gps.pos.alt % 1000 < 750) {
         data.target = targetPresets.south->getTarget(data);  
       } else {
         data.target = targetPresets.west->getTarget(data);
